+++ config/routes.rb
@@ -29,6 +29,12 @@ Rails.application.routes.draw do
 
   authenticate :user, ->(u) { u.admin? } do
     mount Flipper::UI.app(Flipper) => '/admin/flipper'
+
+    namespace :admin do
+      resource :settings, only: %i[show update] do
+        post :test_geocoding
+      end
+    end
   end
 
   # We want to return a nice error message if the user is not authorized to access Sidekiq.
@@ -52,9 +58,6 @@ Rails.application.routes.draw do
 
     resources :background_jobs, only: %i[index create]
     patch 'background_jobs', to: 'background_jobs#update'
-    resource :geocoding, only: %i[show update], controller: 'geocoding' do
-      post :test
-    end
     resource :visits, only: %i[show update]
     resources :users, only: %i[index show create destroy edit update] do
       member do